// FoodTrackerClient/food_tracker/src/ClientGenerator/generated/models/FoodItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Meal Tracker API
 * API documentation for the Meal Tracker Service
 *
 * The version of the OpenAPI document: v1
 * Contact: your.email@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FoodItem
 */
export interface FoodItem {
    /**
     * 
     * @type {string}
     * @memberof FoodItem
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FoodItem
     */
    name?: string | null;
    /**
     * 
     * @type {number}
     * @memberof FoodItem
     */
    calories?: number;
}

/**
 * Check if a given object implements the FoodItem interface.
 */
export function instanceOfFoodItem(value: object): value is FoodItem {
    return true;
}

export function FoodItemFromJSON(json: any): FoodItem {
    return FoodItemFromJSONTyped(json, false);
}

export function FoodItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): FoodItem {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'calories': json['calories'] == null ? undefined : json['calories'],
    };
}

export function FoodItemToJSON(json: any): FoodItem {
    return FoodItemToJSONTyped(json, false);
}

export function FoodItemToJSONTyped(value?: FoodItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'calories': value['calories'],
    };
}